Punto 2 Casi Terminado

+++ objetos2tp1/src/punto2/TarjetaCredito.java
@@ -2,6 +2,6 @@ package punto2;
 
 public interface TarjetaCredito {
 
-	public float calcularDescuento(Pedido pedidoAsociado);
+	public Double calcularDescuento(Pedido pedidoAsociado);
 
 }
